final working version
instead of spending the time to make code to dissect an input roman
numeral, I used the code to convert to roman numerals to make a
comparison loop, checking each number to find the number that matches
the input numeral. It may not be the most resource efficient, but with
the maximun numeral being 4999, it doesn't need to be. I was aiming more
to save lines of code rather than energy.